refactor: change size_t to int for width, height, and filter parameters across multiple files

+++ include/FAST.hpp
@@ -9,7 +9,7 @@ namespace qlm
 	template<pixel_t T>
 	std::vector<KeyPoint<int>> FAST(
 		const Image<ImageFormat::GRAY, T>& in,
-		const size_t arc_length = 9,
+		const int arc_length = 9,
 		const T threshold = 10,
 		const bool non_max_suppression = true);
 }
