+++ tests/source/filters/Test_Gaussian.cpp
@@ -10,7 +10,7 @@ TEST(Test_shakhbat_cv, Gaussian)
 	const bool load_in = in.LoadFromFile(folder_path + "input.jpg");
 
 	// do the operation
-	size_t filter_size = 3;
+	int filter_size = 3;
 	float sigma = 1.0f;
 	t.Start();
 	auto out = qlm::Gaussian<qlm::ImageFormat::RGB, uint8_t>(in, filter_size, sigma);
